enable 64 bit

+++ internal/ceres/compressed_row_sparse_matrix.h
@@ -109,7 +109,7 @@ class CompressedRowSparseMatrix : public SparseMatrix {
   void ToTextFile(FILE* file) const final;
   int num_rows() const final { return num_rows_; }
   int num_cols() const final { return num_cols_; }
-  int num_nonzeros() const final { return rows_[num_rows_]; }
+  int64_t num_nonzeros() const final { return rows_[num_rows_]; }
   const double* values() const final { return &values_[0]; }
   double* mutable_values() final { return &values_[0]; }
